feat(bivariate-matrix-api): 12425 - switch from graphql to new request fromat required by backend for bivariate-matrix and bivariate-color-manager
