+++ src/core/bivariate/index.ts
@@ -1,0 +1,2 @@
+export * from './constants';
+export * from './createBivariateGraphQLQuery';
