--- src/features/bivariate_color_manager/utils/createBivariateColorsGraphQLQuery.ts
@@ -1,49 +1,0 @@
-export function createBivariateColorsGraphQLQuery() {
-  return `
-      fragment AxisFields on Axis {
-        label
-        steps {
-          label
-          value
-        }
-        quality
-        quotient
-      }
-
-    query getPolygonStatistics {
-      polygonStatistic(polygonStatisticRequest:{}) {
-        bivariateStatistic {
-          axis {
-            ...AxisFields
-          }
-          meta {
-            max_zoom
-            min_zoom
-          }
-          indicators {
-            name
-            label
-            direction
-          }
-          correlationRates {
-            x {
-              ...AxisFields
-            }
-            y {
-              ...AxisFields
-            }
-            quality
-            avgCorrelationY
-          }
-          colors {
-            fallback
-            combinations {
-              color
-              corner
-              color_comment
-            }
-          }
-        }
-      }
-    }`;
-}
